requirements.txt
gherkin-official
devtools-ai
absl-py
aenum
ale-py
allure-pytest
allure-python-commons
anyio
Appium-Python-Client
appnope
argon2-cffi
argon2-cffi-bindings
asttokens
astunparse
async-generator
attrs
Babel
backcall
backports.functools-lru-cache
beautifulsoup4
black
bleach
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
coverage
cryptography
cycler
debugpy
decorator
defusedxml
delayed-assert
dictionary
dill
entrypoints
et-xmlfile
execnet
executing
fake-useragent
fastjsonschema
filelock
flatbuffers
fonttools
free-proxy
future
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gym
gym-notices
h11
h5py
huggingface-hub
huggingface-sb3
idna
imageio
imageio-ffmpeg
importlib-metadata
importlib-resources
iniconfig
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
json5
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-limit-output
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
keras
Keras-Preprocessing
kiwisolver
libclang
lxml
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mistune
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
nose
notebook
notebook-shim
numpy
oauthlib
openai
openpyxl
opt-einsum
outcome
packaging
pandas
pandas-stubs
pandocfilters
parso
pathspec
patsy
pexpect
pickle5
pickleshare
Pillow
pip-review
platformdirs
pluggy
polling2
prometheus-client
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
py
pyasn1
pyasn1-modules
pycparser
pyglet
Pygments
pyOpenSSL
pyotp
pyparsing
pyrsistent
PySocks
pytest
pytest-forked
pytest-html
pytest-metadata
pytest-parallel
pytest-reportportal
pytest-xdist
python-dateutil
python-dotenv
python-http-client
python-jsonrpc-server
python-language-server
pytz
PyVirtualDisplay
pywsitest
PyYAML
pyzmq
randomize
regex
reportportal-client
requests
requests-oauthlib
retry
rsa
sacremoses
scikit-learn
scipy
seaborn
selene@git+https://github.com/yashaka/selene.git@6c1bb5f10847c6fc9f0794af05a0d238f2ed7f2f
selenium
self
Send2Trash
sendgrid
singledispatch
six
sniffio
sortedcontainers
soupsieve
stable-baselines3
stack-data
starkbank-ecdsa
statsmodels
tblib
termcolor
terminado
testpath
testrail
threadpoolctl
tinycss2
tokenizers
toml
tomli
torch
tornado
tqdm
traitlets
transformers
trio
trio-websocket
types-pytz
typing_extensions
ujson
unsync
urllib3
wasabi
wcwidth
webdriver-manager
webencodings
websocket-client
websockets
Werkzeug
widgetsnbextension
wrapt
wsproto
zipp
selen@git+https://github.com/SatoDeNoor/selen.git
parver
shortuuid
pytest-bdd
pytest-json-report
allure-combine
